refactor: remove SmallCard component in favor of existing card implementations
